+++ GlobalConfigs.h
@@ -35,7 +35,7 @@ struct ColorComponent {
 struct GlobalConfig {
   int fontSize;
   float wordHeight, wordWidth;
-  int breakPointMarkWidth, blockFoldingMarkWidth;
+  float breakPointMarkWidth, blockFoldingMarkWidth;
   int lineNumberThresholdWidth;
 
   int cursorBlinkTimeInSeconds;
@@ -55,8 +55,8 @@ struct GlobalConfig {
   void setFontSize (int);
   void setWordHeight (float);
   void setWordWidth (float);
-  void setBreakPointMarkWidth (int);
-  void setBlockFoldingMarkWidth (int);
+  void setBreakPointMarkWidth (float);
+  void setBlockFoldingMarkWidth (float);
   void setLineNumberThresholdWidth (int);
   void setCursorBlinkTimeInSeconds (int);
   void setCursorHeight (float);
@@ -71,8 +71,8 @@ struct GlobalConfig {
   int getFontSize ();
   float getWordHeight ();
   float getWordWidth ();
-  int getBreakPointMarkWidth ();
-  int getBlockFoldingMarkWidth ();
+  float getBreakPointMarkWidth ();
+  float getBlockFoldingMarkWidth ();
   int getLineNumberThresholdWidth ();
   int getCursorBlinkTimeInSeconds ();
   float getCursorHeight ();
